pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "azure-core>=1.34.0",
    "azure-identity>=1.23.0",
    "openai>=1.87.0",
    "opencv-python>=4.11.0.86",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "pymupdf>=1.26.1",
    "pyodbc>=5.2.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
